* Updated README.md  * Introduced INDI version macro to handle different libindi versions  * Extracted to from findFoxusStar() to new SimpleStarDetector class

+++ source/focus_finder/common/focus_finder_profile.cpp
@@ -204,7 +204,7 @@ void FocusFinderProfileT::setNumFramesPerFocusPos(size_t numFramesPerFocusPos) {
     mNumFramesPerFocusPos = numFramesPerFocusPos;
 }
 
-size_t FocusFinderProfileT::getStarDetectionSnrBoundary() const {
+float FocusFinderProfileT::getStarDetectionSnrBoundary() const {
     return mStarDetectionSnrBoundary;
 }
