+++ source/focus_finder/common/include/focus_finder_profile.h
@@ -123,7 +123,7 @@ public:
 
     void setNumFramesPerFocusPos(size_t numFramesPerFocusPos);
 
-    [[nodiscard]] size_t getStarDetectionSnrBoundary() const;
+    [[nodiscard]] float getStarDetectionSnrBoundary() const;
 
     void setStarDetectionSnrBoundary(float starDetectionSnrBoundary);
